Start of syncing new snippets

+++ src/snipline_cli/services/temp_snippet_editor_file.cr
@@ -15,11 +15,14 @@ This section supports **Markdown**
 """
 is_pinned = true
 snippet_alias = ""
+sync_to_cloud = true
 >
 
     def create
       config = SniplineCli.config
-      File.write(File.expand_path("#{config.get("general.temp_dir")}/temp.toml"), TEMPLATE)
+      unless File.exists?(File.expand_path("#{config.get("general.temp_dir")}/temp.toml"))
+        File.write(File.expand_path("#{config.get("general.temp_dir")}/temp.toml"), TEMPLATE)
+      end
     end
 
     def read
@@ -33,6 +36,13 @@ snippet_alias = ""
         snippet_alias: toml["snippet_alias"].as(String),
         tags: [] of String
       )
+    end
+
+    def delete
+      config = SniplineCli.config
+      unless File.exists?(File.expand_path("#{config.get("general.temp_dir")}/temp.toml"))
+        File.delete(File.expand_path("#{config.get("general.temp_dir")}/temp.toml"))
+      end
     end
   end
 end
